[29-0012] Nuevo pre-diseno de los habitos
